Project 2 - Covid-19 vs. Influenza Cases
As a laboratory assistant, my coworkers and I noticed that there were a sudden decrease in influenza cases. I compared Covid-19 deaths to Influenza deaths as well as the activity level to determine if masks were the reason for it. The code in comments are code that is not used for the final tableau project.